+++ constants.py
@@ -1,0 +1,13 @@
+import os
+
+
+FPS = 60
+
+size = WIDTH, HEIGHT = 500, 500
+tile_width = tile_height = 50
+
+ROOT_PATH = os.path.dirname(os.path.abspath(__file__))
+
+GRAVITY = 0.35
+JMP_POWER = 10
+MOVE_SPEED = 7
