+++ groups.py
@@ -1,0 +1,5 @@
+from pygame import sprite as s
+
+all_sprites = s.Group()
+tiles_group = s.Group()
+player_group = s.Group()
